+++ config/alerts.yml
@@ -5,7 +5,7 @@
 
 sources:
   conditions:
-    class: Alerts::Sources::ConditionsSource
+    class: Alerts::ConditionsSource
     fields:
       am_min_height:
         type: Integer
@@ -47,7 +47,7 @@ sources:
           - GOOD_TO_EPIC
           - EPIC
   wave:
-    class: Alerts::Sources::WaveSource
+    class: Alerts::WaveSource
     fields:
       direction:
         type: Degree
