+++ lib/tasks/alerts.rake
@@ -38,7 +38,7 @@ end
 def generate_and_send_email(user)
   Rails.logger.info "alerts:send_emails - Generating email for User##{user.id}"
 
-  alert_ids_and_days_of_week = Alerts::Emails::DigestDataGenerator.new(user.id).generate
+  alert_ids_and_days_of_week = Alerts::DigestDataGenerator.new(user.id).generate
 
   return unless alert_ids_and_days_of_week.any?
